Add files via upload
This project, developed using the powerful trio of HTML, CSS, and JavaScript, is a state-of-the-art tool that generates highly secure passwords with just a click. With the increasing need for robust security measures in our digital landscape, this Password Generator comes as a timely solution, ensuring that your online presence remains safeguarded against any potential threats.

Key features of the Password Generator include:

✅ Customizable password length for added flexibility .
✅ Advanced algorithm ensuring the highest level of password complexity.
✅ User-friendly interface for a seamless experience.
✅ checkboxes for applying the strength of the password you want.
✅ You can copy the  the password after the generating 

Feel free to check out the live demo of the Password Generator  and experience the ease of generating strong, secure passwords effortlessly.